minor build fix for Linux

+++ renderdoc/maths/formatpacking.h
@@ -1,18 +1,18 @@
 /******************************************************************************
  * The MIT License (MIT)
- * 
+ *
  * Copyright (c) 2014 Crytek
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -59,9 +59,9 @@ inline Vec3f ConvertFromR11G11B10(uint32_t data)
 		(data>>22) & 0x1f,
 	};
 	int32_t exponents[3] = {
-		(data>> 6) & 0x1f,
-		(data>>17) & 0x1f,
-		(data>>27) & 0x1f,
+		int32_t(data>> 6) & 0x1f,
+		int32_t(data>>17) & 0x1f,
+		int32_t(data>>27) & 0x1f,
 	};
 
 	Vec3f ret;
